Avoid implementation.rs
Use more descriptive names or move into `mod.rs` where appropriate
