+++ opcua-client/src/session/event_loop.rs
@@ -17,9 +17,8 @@ use opcua_types::{
 
 use super::{
     connect::{SessionConnectMode, SessionConnector},
-    implementation::SessionState,
     services::subscriptions::event_loop::{SubscriptionActivity, SubscriptionEventLoop},
-    Session,
+    Session, SessionState,
 };
 
 /// A list of possible events that happens while polling the session.
